Kubernetes support
The host docker connector now accounts for the fact that only the Pause/POD container in a Kubernetes Pod has the IP address for the pod as a whole. Since Amalgam8 refers to the service instances by IP address, the plugin was attaching to the Pause/POD containers rather than the application containers. The host docker connector now copies this IP address to the neighboring containers inside each of the pods before sending the Container ID/IP address pairs to the portion of the A8 Routing Plugin inside of the container